Closes #3092: Split DCIM models into separate files for easier management
